+++ ovp_users/models/password_recovery.py
@@ -13,7 +13,11 @@ class PasswordRecoveryToken(models.Model):
   def save(self, *args, **kwargs):
     if not self.pk:
       self.token = uuid.uuid4()
-      self.user.mailing().sendRecoveryToken({'token': self})
+
+      if self.user.exceeded_login_attempts:
+        self.user.mailing().sendExceededLoginAttempts({'token': self})
+      else:
+        self.user.mailing().sendRecoveryToken({'token': self})
 
     super(PasswordRecoveryToken, self).save(*args, **kwargs)