+++ CHANGES.rst
@@ -202,3 +202,4 @@ v1.1.9[unreleased]
 * Fix for case insensitive user search
 * Add email on LongUserPublicRetrieveSerializer
 * Add EXPIRE_PASSWORD_IN setting
+* Limit login attempts